Day 14, part 2
This one caused me a lot of time and trouble.
I was pretty sure my naive way of part 1 would not hold up much more
than 10 steps and that the answer was probably better searched in
counting the pairs/letters directly, but I could not get my head around
it 100%, resulting in a couple of failed attempts.
Luckily I pingponged my findings with @Jelle-S which set me on the right
track; basically I didn't catch the fact that the pair counting
(`prefq`) needed to be restarted every step and just the counts had to
be copied over.

Then I had a working solution, that is to say: it worked for the sample
data, both 10 and 40 steps, but it did not for the real input data...
this was because I forgot that splitting up the template in pairs could
result in a pair being present more than once.